reorganize custom_library module
