+++ source/extensions/filters/common/expr/evaluator.h
@@ -5,7 +5,6 @@
 #include "source/common/http/headers.h"
 #include "source/common/protobuf/protobuf.h"
 #include "source/extensions/filters/common/expr/context.h"
-#include "source/extensions/filters/common/expr/library/custom_functions.h"
 #include "source/extensions/filters/common/expr/library/custom_library.h"
 
 #include "eval/public/activation.h"
@@ -27,8 +26,6 @@ using Expression = google::api::expr::runtime::CelExpression;
 using ExpressionPtr = std::unique_ptr<Expression>;
 
 using CelValue = google::api::expr::runtime::CelValue;
-using ConstCelFunction = Envoy::Extensions::Filters::Common::Expr::Library::ConstCelFunction;
-using CustomVocabularyWrapper = Envoy::Extensions::Filters::Common::Expr::Library::CustomVocabularyWrapper;
 using CustomLibrary = Envoy::Extensions::Filters::Common::Expr::Library::CustomLibrary;
 
 // Creates an activation providing the common context attributes.
@@ -43,7 +40,7 @@ ActivationPtr createActivation(Protobuf::Arena& arena, const StreamInfo::StreamI
 // for intermediate evaluation results.
 // Throws an exception if fails to construct an expression builder.
 BuilderPtr createBuilder(Protobuf::Arena* arena,
-                               const CustomLibrary* custom_library);
+                         const CustomLibrary* custom_library);
 
 // Creates an interpretable expression from a protobuf representation.
 // Throws an exception if fails to construct a runtime expression.
@@ -56,7 +53,7 @@ absl::optional<CelValue> evaluate(const Expression& expr, Protobuf::Arena& arena
                                   const Http::RequestHeaderMap* request_headers,
                                   const Http::ResponseHeaderMap* response_headers,
                                   const Http::ResponseTrailerMap* response_trailers,
-                               const CustomLibrary* custom_library);
+                                  const CustomLibrary* custom_library);
 
 // Evaluates an expression and returns true if the expression evaluates to "true".
 // Returns false if the expression fails to evaluate.
